Converted to .NET 5.0, fixed GetDataUse methods

+++ Api/Entities/DataUse.cs
@@ -8,16 +8,16 @@
         /// <summary>
         /// Total amount of data coming in to pfSense in MB
         /// </summary>
-        public int? In { get; set; }
+        public decimal? In { get; set; }
 
         /// <summary>
         /// Total amount of data going out of pfSense in MB
         /// </summary>
-        public int? Out { get; set; }
+        public decimal? Out { get; set; }
 
         /// <summary>
         /// Total amount of data coming in to and going out of pfSense in MB
         /// </summary>
-        public int? Total { get; set; }
+        public decimal? Total { get; set; }
     }
 }
